Add files via upload
This Waveshaper needs an SPI Oled and an Encoder connected to the patch.Init() and an SDcard for saving / loading waveshapes. 
Navigating with the encoder : turn and click to validate.
EDT : Edit Menu, allows to change the inflexion points for the waveshape. 16 points are available.
OSC : Sine / Tri/ Saw and input on Left are available. For SINE/TRI and SAW,  3 detunable osc play at the same time ( detune amount is managed by pot CV1)
CV pitch is received in CV6
CV5 can receive an LFO or an enveloppe to modulate the amplitude of the initial oscillator.  CV5 is attenuated by the Pot C2

If no sound at start : enter EDT menu and change one inflexion point... and send a pitch into CV6.